refactor(mobile): split ChatPage into chat/ modules with tested models
ChatPage.tsx mixed seven components, a hook, a 240-line markdown style
tree rebuilt per bubble, and untested pure logic in one 1500-line file.
It now keeps only page wiring (302 lines) over src/agent/chat/:

- chat-markdown.ts: the markdown style tree as a module factory
  memoized per palette, so chat rows share two frozen trees instead of
  building one per instance
- chat-message-model.ts: isFinalMarkdownNode, isSameCalendarDay,
  chatDateLabel, and the pure resolveMarkdownLink target resolution,
  covered by 25 table-driven vitest cases; the Linking/WebBrowser side
  effect stays at the call site
- chat-event.tsx, chat-composer.tsx, scroll-to-bottom-button.tsx,
  use-transcript-word-haptics.ts: moved subcomponents, JSX and styles
  unchanged
- chat-transcript.tsx: the FlatList and its overlays as a memoized
  child fed only keystroke-stable props, so typing and per-word voice
  transcripts re-render the composer row alone instead of the whole
  transcript with fresh inline style arrays and footer elements

MessageContextMenu is generic over its action id, which deletes the
"id as MessageActionId" cast at the call site, and reads
process.env.EXPO_OS like the rest of the app. The per-agent composer
hold wiring is untouched.
